pyproject.toml
[project]
name = "visxgenai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "altair==5.5.0",
    "draco",
    "dspy==2.6.27",
    "duckdb==1.3.2",
    "fastapi==0.116.1",
    "google-cloud-aiplatform==1.108.0",
    "hrid==0.3.0",
    "httpx==0.28.1",
    "infisicalsdk==1.0.9",
    "langfuse==3.2.2",
    "litellm==1.75.5.post1",
    "markdownify==1.2.0",
    "nest-asyncio==1.6.0",
    "numpy==2.1.3",
    "obstore==0.7.3",
    "openai==1.99.9",
    "openinference-instrumentation-dspy==0.1.27",
    "orjson==3.11.1",
    "pillow==11.3.0",
    "polars==1.32.0",
    "pydantic==2.11.7",
    "sqlglot==27.6.0",
    "typing-extensions==4.14.1",
    "vl-convert-python==1.8.0",
]

[dependency-groups]
dev = [
    "marimo[recommended]>=0.15.1",
    "python-lsp-server>=1.12.2",
    "ruff>=0.12.0",
]

[tool.marimo.runtime]
dotenv = [".env"]
pythonpath = ["src"]
output_max_bytes = 10_000_000_000

[tool.uv.sources]
# Custom Draco build from https://github.com/cmudig/draco2/pull/949 using tuned rules & weights and more convenient API
draco = { url = "https://visxgenai-cdn.peter.gy/pip/draco/draco-2.0.1-py3-none-any.whl" }

[tool.ruff]
line-length = 88
lint.select = [
    "E",
    "F",
    "W",
    # isort
    "I001",
]
lint.ignore = ["E501", "W191"]
src = ["src"]
fix = true
